# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AL-RAHMAN TRADERS Ledger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sale-orders:
    get:
      operationId: listSaleOrders
      tags: [saleOrders]
      summary: List all sale orders
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: includeReversed
          in: query
          description: Include reversed originals and reversal rows (default excludes them — see the correction workflow).
          schema:
            type: boolean
      responses:
        "200":
          description: Sale order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleOrder"
    post:
      operationId: createSaleOrder
      tags: [saleOrders]
      summary: Create a new sale order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleOrderInput"
      responses:
        "201":
          description: Sale order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleOrder"
  /sale-orders/{id}:
    get:
      operationId: getSaleOrder
      tags: [saleOrders]
      summary: Get a sale order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleOrder"
  /sale-orders/{id}/correct:
    post:
      operationId: correctSaleOrder
      tags: [saleOrders]
      summary: Correct (or void) a posted sale order — never edits or deletes it in place
      description: >
        Standard correction workflow: the original row is never mutated (only its status
        flips to 'reversed') or deleted. This posts a 'reversal' row that cancels it, and
        — unless `void` is set — a new 'posted' row with the corrected data. Fails with
        409 if the target row isn't currently 'posted' (already reversed/corrected).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleOrderCorrectionInput"
      responses:
        "200":
          description: Correction applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleOrderCorrectionResult"
        "409":
          description: The target row is not currently posted

  # ── SALE RETURNS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SaleOrderItem:
      type: object
      required: [id, productId, productName, qty, rate, amount]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
        notes:
          type: ["string", "null"]
        costPrice:
          type: ["number", "null"]
    SaleOrder:
      type: object
      required: [id, customerId, customerName, date, totalAmount, discountAmount, netAmount, createdAt, items, status]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        date:
          type: string
          format: date
        vehicleNo:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        billtyNo:
          type: ["string", "null"]
        totalAmount:
          type: number
        discountAmount:
          type: number
          description: Flat discount agreed at sale time. Never exceeds totalAmount.
        netAmount:
          type: number
          description: totalAmount - discountAmount — what the customer actually owes for this order.
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleOrderItem"
        status:
          type: string
          enum: [posted, reversed, reversal]
        reversesId:
          type: ["integer", "null"]
        correctsId:
          type: ["integer", "null"]
    SaleOrderItemInput:
      type: object
      required: [productId, qty]
      properties:
        productId:
          type: integer
        qty:
          type: number
        rate:
          type: number
        notes:
          type: string
    SaleOrderInput:
      type: object
      required: [customerId, date, items]
      properties:
        customerId:
          type: integer
        date:
          type: string
          format: date
        vehicleNo:
          type: string
        driverName:
          type: string
        billtyNo:
          type: string
        notes:
          type: string
        discountAmount:
          type: number
          description: Flat discount off the order total. Must not exceed the sum of item amounts.
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleOrderItemInput"
    SaleOrderCorrectionInput:
      type: object
      description: >
        Either set void=true to reverse the original with no replacement, or omit it and
        supply the corrected data (same shape as SaleOrderInput) to reverse-and-replace.
      properties:
        void:
          type: boolean
        reason:
          type: string
        customerId:
          type: integer
        date:
          type: string
          format: date
        vehicleNo:
          type: string
        driverName:
          type: string
        billtyNo:
          type: string
        notes:
          type: string
        discountAmount:
          type: number
          description: Flat discount off the order total. Must not exceed the sum of item amounts.
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleOrderItemInput"
    SaleOrderCorrectionResult:
      type: object
      required: [original, reversal]
      properties:
        original:
          $ref: "#/components/schemas/SaleOrder"
        reversal:
          $ref: "#/components/schemas/SaleOrder"
        correction:
          $ref: "#/components/schemas/SaleOrder"
...